Code clean-up. Some for-loop and cast cleanup. No functional change.

+++ src/com/digero/common/abctomidi/AbcToMidi.java
@@ -988,7 +988,7 @@ public class AbcToMidi
 								track.add(noteOff);
 								noteOffEvents.add(noteOff);
 
-								tiedNotes.remove((Integer) noteId);
+								tiedNotes.remove(noteId);
 							}
 						}
