+++ src/com/digero/common/view/VolumeBar.java
@@ -48,8 +48,7 @@ public class VolumeBar extends JPanel
 		Graphics2D g2 = (Graphics2D) g;
 		g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
 
-		int ptrPos = (int) (SIDE_PAD + (getWidth() - 2 * SIDE_PAD) * volumizer.getVolume()
-				/ VolumeTransceiver.MAX_VOLUME);
+		int ptrPos = SIDE_PAD + (getWidth() - 2 * SIDE_PAD) * volumizer.getVolume() / VolumeTransceiver.MAX_VOLUME;
 
 		final int x = 0;
 		final int y = (PTR_HEIGHT - BAR_HEIGHT) / 2;
@@ -97,14 +96,14 @@ public class VolumeBar extends JPanel
 	{
 		private int getPosition(int x)
 		{
-			int pos = (int) ((x + 1 - SIDE_PAD) * VolumeTransceiver.MAX_VOLUME / (getWidth() - 2 * SIDE_PAD));
+			int pos = (x + 1 - SIDE_PAD) * VolumeTransceiver.MAX_VOLUME / (getWidth() - 2 * SIDE_PAD);
 			if (pos < 0)
 			{
 				pos = 0;
 			}
 			if (pos > VolumeTransceiver.MAX_VOLUME)
 			{
-				pos = (int) VolumeTransceiver.MAX_VOLUME;
+				pos = VolumeTransceiver.MAX_VOLUME;
 			}
 			return pos;
 		}
